chore: chiseling down spec.
